step3, main, controller 클래스 분리 및 경로 정리
