+++ src/main/kotlin/step3/racing/model/RandomGenerator.kt
@@ -1,4 +1,4 @@
-package step3.model
+package step3.racing.model
 
 interface RandomGenerator {
     fun generator(range: IntRange): Int
