+++ src/test/kotlin/study/racing/InputViewTest.kt
@@ -3,7 +3,7 @@ package study.racing
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import step3.view.InputViewFactory
+import step3.racing.view.InputViewFactory
 
 class InputViewTest {
     @Test
